players added to corners sequentially, bomb smaller, bombs power up working
