+++ backend/server/makeBoard.js
@@ -2,7 +2,7 @@
 export function AddBricks() {
     var layout = [
     [[1], [1], [1], [1], [1], [1], [1], [1], [1], [1], [1], [1], [1], [1], [1]],
-    [[1], [3], [0], [0], [0], [0], [0], [0], [0], [0], [0], [0], [0], [4], [1]], 
+    [[1], [0], [0], [0], [0], [0], [0], [0], [0], [0], [0], [0], [0], [0], [1]], 
     [[1], [0], [1], [0], [1], [0], [1], [0], [1], [0], [1], [0], [1], [0], [1]], 
     [[1], [0], [0], [0], [0], [0], [0], [0], [0], [0], [0], [0], [0], [0], [1]],
     [[1], [0], [1], [0], [1], [0], [1], [0], [1], [0], [1], [0], [1], [0], [1]],
@@ -14,7 +14,7 @@ export function AddBricks() {
     [[1], [0], [1], [0], [1], [0], [1], [0], [1], [0], [1], [0], [1], [0], [1]],
     [[1], [0], [0], [0], [0], [0], [0], [0], [0], [0], [0], [0], [0], [0], [1]],
     [[1], [0], [1], [0], [1], [0], [1], [0], [1], [0], [1], [0], [1], [0], [1]],
-    [[1], [5], [0], [0], [0], [0], [0], [0], [0], [0], [0], [0], [0], [6], [1]],
+    [[1], [0], [0], [0], [0], [0], [0], [0], [0], [0], [0], [0], [0], [0], [1]],
     [[1], [1], [1], [1], [1], [1], [1], [1], [1], [1], [1], [1], [1], [1], [1]],
 ]
     for (let row = 0; row < layout.length; row++) {
