+++ frontend/components/bomb.jsx
@@ -1,26 +1,9 @@
-import { ctx, tileMap, drawPlayer, players } from "./game"
-import { self } from "./game";
-import deadAngel from "../static/sounds/deadAngel.ogg"
-import { whoAmI } from "./waitForPlayers";
+import { ctx, tileMap, players, bombs, self } from "./game"
+import { drawPlayer } from "./move";
+
+
 //import {whoAmI} from "./waitForPlayers"
 
-//=======> Start Heart and Explosion variables and functions <=======
-//var isPlaying = false;
-var tries = 0;
-var lostHeart;
-var explosion;
-var audio4 = new Audio(deadAngel);
-
-//sets off explosion once after loss of one life, and plays 'deadAngel'
-function puff(t) {
-    audio4.currentTime = 0;
-    audio4.play();
-    explosion = document.getElementById(`explosion${t}`);
-    explosion.style.opacity = "1";
-    setTimeout(() => {
-        explosion.style.opacity = "0"
-    }, 500);
-}
 
 //=======> End Heart and Explosion variables and functions <=======
 
@@ -93,121 +76,7 @@ function drawExplosion(b) {
 }
 
 
-function undrawExplosion(b) {
-    let o = [20, -20, 0, 0]
-    let p = [0, 0, 10, -10]
-  
-    let check = [b.col+b.count, b.col-b.count, b.row+b.count, b.row-b.count]
-    let bcheck = [b.col+b.count-1, b.col-b.count-1, b.row+b.count-1, b.row-b.count-1]
-for (let i=0; i<4; i++) {
-    console.log("remove tile\ncount", b.count, " row, col:", b.row, b.col)
-    if (check[i]<14 && check[i]>0){
-        console.log("position within tilemap")
-        let spot 
-        let pSpot
-        let row 
-        let col
-        if (i<2){
-           row = b.row 
-           col = check[i]
-            spot = tileMap.map[b.row][check[i]]
-            pSpot = bcheck[i]>0? tileMap.map[b.row][bcheck[i]] : []
-            console.log("spot", spot, "pSpot", pSpot)
-        } else {
-            row = check[i]
-            col = b.col
-            spot = tileMap.map[check[i]][b.col]
-            pSpot = bcheck[i]>0? tileMap.map[bcheck[i]][b.col] : []
-            console.log("spot", spot, "pSpot", pSpot)
-        }
-        console.log("can I unexplode?\n is there a wall here?", spot.includes(1), "\ndoes the previous spot have a wall?", pSpot.includes(1), "\nis this the central sqare?", b.count)
-        if ((!spot.includes(1) && !pSpot.includes(1)) || b.count === 0) {
-            console.log("no wall present, drawing explosion")
-                ctx.drawImage(
-                    tileMap.grass,
-                    0,
-                    0,
-                    779,
-                    779,
-                    b.cX + (o[i] * b.count),
-                    b.cY + (p[i] * b.count),
-                    20,
-                    10
-                )
 
-                spot.forEach(
-                    (n) => {
-                        console.log("what is n, while unexploding?", n)
-                        if (n > 2 && n < 7) {
-                            //tries++
-                            console.log("Exploded player index is:", n - 3)
-                            console.log("tries at start, 2 x player's index, and player's lives", tries, self.index, players[n - 3].index, players[n - 3].lives)
-                            //check players number of lives
-                            if (players[n - 3].lives > 1) {
-                                if (players[n - 3].index === self.index) {
-                                    //derive which heart to remove
-                                    console.log("type of players.index", typeof (players[n - 3].index))
-
-                                    //Bomberman looses one life
-                                    players[n - 3].lives--
-                                    //determine number of tries
-                                    tries = 3 - players[n - 3].lives
-                                    //remove heart that has the same number as the number of tries
-                                    lostHeart = document.getElementById(`heart${tries}`)
-                                    //heart exploding and disappearing
-                                    lostHeart.style.opacity = "0"
-                                    puff(tries);
-
-                                }
-
-                                drawPlayer(n - 3)
-
-                            } else {
-                                //if 1: take away life and show game over
-                                if (players[n - 3].index === self.index) {
-                                    tries = 3;
-                                    lostHeart = document.getElementById(`heart${tries}`)
-                                    lostHeart.style.opacity = "0"
-                                    puff(tries);
-                                    //display end of game message
-                                    document.querySelector("#whoAmI").innerHTML = '<p> GAME OVER<br>'
-                                        + `Better luck next time ${whoAmI}</p>`
-                                        ;
-                                }
-
-                            }
-
-                        }
-                        if (n>7){
-                            console.log("found!!!! drawing power up")
-                            let pX
-                            if (n===8){pX = b.bombs} else if(n===9) {pX = b.flames} else if (n===10){pX = b.speed}
-                            ctx.drawImage(
-                                b.powerUp,
-                                pX,
-                                0,
-                                125,
-                                125,
-                                col*20,
-                                row*10,
-                                20,
-                                10
-                            )
-                        } 
-                    }
-                )
-            }
-            if (b.count === 0) {
-                i = 4
-                spot.splice(0, 1)
-            }
-
-        }
-       
-    }
-    b.count--
-
-}
 
 export function progBomb(b, timeStamp) {
 
@@ -226,30 +95,14 @@ export function progBomb(b, timeStamp) {
         //check to see if explosion is complete > status = exploded
         //flames power up lets explosion reach 1 square further
         let power = players[b.playerI].flames
-        console.log("how many flames power ups are being used")
-        if ((power === 0 && b.count===3) || (power > 0  && b.count=== 3 + power) ) {
+        console.log("how many flames power ups are being used", players[b.playerI].flames)
+        if ((power === 0 && b.count===2) || (power > 0  && b.count=== 2 + power) ) {
         //if( b.count > 2){
             console.log("bomb exploded")
-            b.status = "exploded"
+            b.status = "bomb exploded"
             b.count--
             players[b.playerI].flames = 0
         }
     
 }
-export function unBomb(b, timeStamp) {
-    const increment = timeStamp - b.previousTimeStamp
-    //make sure 3 secs have passed before starting  explosion
-    if (increment >= 400) {
-        // progressively display explosion
-        //X++54 Y++50.6  x-379 y-354
-        console.log("de-explosion starting")
-        undrawExplosion(b)
-        b.previousTimeStamp = timeStamp;
-    }
-    //check to see if de-explosion is complete > status = complete 
-    if (b.count < 0) {
-        console.log("explosion complete")
-        b.status = "complete"
-    }
 
-}
